feat: add types for post
