+++ news-api/src/entities/post/model/index.ts
@@ -1,1 +1,2 @@
 export { selectAllPosts, postSlice, postSelectors } from './slice';
+export type { Post, Filters } from './types.ts';
