+++ news-api/src/entities/post/model/types.ts
@@ -1,0 +1,12 @@
+export interface Post {
+    id: number;
+    userId: number;
+    title: string;
+    body: string;
+}
+
+export interface Filters {
+    limit?: number;
+    userId?: number;
+    page?: number;
+}
